requirements.txt
aiofiles
aiohttp
aiolimiter
aioshutil
dataclasses_json
dnspython
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
httpx
m3u8
mutagen
natsort
pathvalidate
Pillow
psycopg2-binary
psutil
pycryptodome
pycryptodomex
pymongo
pyrogram
python-dotenv
python-magic
requests
tenacity
tgcrypto
tidal-dl-ng
tidalapi
toml
typer
